// catalog-rest-service/src/main/resources/ui/src/generated/type/entityLineage.ts
/* eslint-disable @typescript-eslint/no-explicit-any */
/**
 * Licensed to the Apache Software Foundation (ASF) under one or more
 * contributor license agreements. See the NOTICE file distributed with
 * this work for additional information regarding copyright ownership.
 * The ASF licenses this file to You under the Apache License, Version 2.0
 * (the "License"); you may not use this file except in compliance with
 * the License. You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

/**
 * This schema defines the type used for lineage of an entity.
 */
export interface EntityLineage {
  downstreamEdges?: Edge[];
  /**
   * Primary entity for which this lineage graph is created.
   */
  entity: EntityReference;
  nodes?: EntityReference[];
  upstreamEdges?: Edge[];
}

/**
 * Edge in the lineage graph from one entity to another by entity IDs.
 */
export interface Edge {
  description?: string;
  /**
   * From entity that is upstream of lineage edge.
   */
  fromEntity?: string;
  /**
   * To entity that is downstream of lineage edge.
   */
  toEntity?: string;
}

/**
 * Primary entity for which this lineage graph is created.
 *
 * This schema defines the EntityReference type used for referencing an entity.
 * EntityReference is used for capturing relationships from one entity to another. For
 * example, a table has an attribute called database of type EntityReference that captures
 * the relationship of a table `belongs to a` database.
 */
export interface EntityReference {
  /**
   * Optional description of entity.
   */
  description?: string;
  /**
   * Link to the entity resource.
   */
  href?: string;
  /**
   * Unique identifier that identifies an entity instance.
   */
  id: string;
  /**
   * Name of the entity instance. For entities such as tables, databases where the name is not
   * unique, fullyQualifiedName is returned in this field.
   */
  name?: string;
  /**
   * Entity type/class name - Examples: `database`, `table`, `metrics`, `redshift`, `mysql`,
   * `bigquery`, `snowflake`...
   */
  type: string;
}